times test: Avoid gcc warnings on Linux/x32.
* tests/test-times.c (main): Really cast printf arguments from clock_t
to 'long int'.

+++ tests/test-times.c
@@ -61,10 +61,10 @@ main (int argc, char *argv[])
       printf ("clk_tck %ld\n", (long int) clk_tck);
 
       printf ("t %ld\n", (long int) t);
-      printf ("tms.tms_utime %ldms\n", ((long int) tms.tms_utime * 1000) / clk_tck);
-      printf ("tms.tms_stime %ldms\n", ((long int) tms.tms_stime * 1000) / clk_tck);
-      printf ("tms.tms_cutime %ldms\n", ((long int) tms.tms_cutime * 1000) / clk_tck);
-      printf ("tms.tms_cstime %ldms\n", ((long int) tms.tms_cstime * 1000) / clk_tck);
+      printf ("tms.tms_utime %ldms\n", ((long int) tms.tms_utime * 1000) / (long int) clk_tck);
+      printf ("tms.tms_stime %ldms\n", ((long int) tms.tms_stime * 1000) / (long int) clk_tck);
+      printf ("tms.tms_cutime %ldms\n", ((long int) tms.tms_cutime * 1000) / (long int) clk_tck);
+      printf ("tms.tms_cstime %ldms\n", ((long int) tms.tms_cstime * 1000) / (long int) clk_tck);
     }
 
   if (argc > 1)
@@ -95,10 +95,10 @@ main (int argc, char *argv[])
       printf ("clk_tck %ld\n", (long int) clk_tck);
 
       printf ("t %ld\n", (long int) t);
-      printf ("tms.tms_utime %ldms\n", ((long int) tms.tms_utime * 1000) / clk_tck);
-      printf ("tms.tms_stime %ldms\n", ((long int) tms.tms_stime * 1000) / clk_tck);
-      printf ("tms.tms_cutime %ldms\n", ((long int) tms.tms_cutime * 1000) / clk_tck);
-      printf ("tms.tms_cstime %ldms\n", ((long int) tms.tms_cstime * 1000) / clk_tck);
+      printf ("tms.tms_utime %ldms\n", ((long int) tms.tms_utime * 1000) / (long int) clk_tck);
+      printf ("tms.tms_stime %ldms\n", ((long int) tms.tms_stime * 1000) / (long int) clk_tck);
+      printf ("tms.tms_cutime %ldms\n", ((long int) tms.tms_cutime * 1000) / (long int) clk_tck);
+      printf ("tms.tms_cstime %ldms\n", ((long int) tms.tms_cstime * 1000) / (long int) clk_tck);
     }
 
   return 0;
